requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
Flask-WTF==1.2.1 